Shove yet more readline logic into readline.c
