+++ stdenv.h
@@ -111,6 +111,7 @@ extern void *qsort(
 
 #if HAVE_READLINE
 # include <stdio.h>
+# include <readline/readline.h> /* ugly but we need it in both input.c and readline.c */
 #endif
 
 #include <sys/wait.h>
